# openapi.json
{"openapi":"3.1.0","info":{"title":"Wellfounded API","version":"0.1.0"},"servers":[{"url":"http://localhost:8000","description":"Local dev"}],"paths":{"/orgs/admin/users":{"get":{"tags":["orgs"],"summary":"List Org Users Stub","operationId":"list_org_users_stub_orgs_admin_users_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"items":{"additionalProperties":{"type":"string"},"type":"object"},"type":"array"},"type":"object","title":"Response List Org Users Stub Orgs Admin Users Get"}}}}}}},"/cases":{"post":{"tags":["cases"],"summary":"Create Case","operationId":"create_case_cases_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaseCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaseDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["cases"],"summary":"List Cases","operationId":"list_cases_cases_get","parameters":[{"name":"status","in":"query","required":false,"schema":{"$ref":"#/components/schemas/CaseQueryStatus","default":"active"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CaseListItem"},"title":"Response List Cases Cases Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases/{case_id}":{"get":{"tags":["cases"],"summary":"Get Case","operationId":"get_case_cases__case_id__get","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Case Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaseDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["cases"],"summary":"Patch Case","operationId":"patch_case_cases__case_id__patch","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Case Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaseUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaseDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases/{case_id}/archive":{"post":{"tags":["cases"],"summary":"Archive Case","operationId":"archive_case_cases__case_id__archive_post","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Case Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaseDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases/{case_id}/assignments":{"post":{"tags":["cases"],"summary":"Patch Assignments","operationId":"patch_assignments_cases__case_id__assignments_post","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Case Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaseAssignmentsPatch"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaseDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases/{case_id}/soft-delete":{"post":{"tags":["cases"],"summary":"Soft Delete Case","operationId":"soft_delete_case_cases__case_id__soft_delete_post","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Case Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases/{case_id}/restore":{"post":{"tags":["cases"],"summary":"Restore Case","operationId":"restore_case_cases__case_id__restore_post","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Case Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases/{case_id}/country-conditions":{"post":{"tags":["cases"],"summary":"Request Country Conditions Memo","operationId":"request_country_conditions_memo_cases__case_id__country_conditions_post","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Case Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CountryConditionsGenerateRequest"}}}},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CountryConditionsGenerateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["cases"],"summary":"List Country Conditions Memos","operationId":"list_country_conditions_memos_cases__case_id__country_conditions_get","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Case Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryConditionsMemoSummary"},"title":"Response List Country Conditions Memos Cases  Case Id  Country Conditions Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases/{case_id}/country-conditions/{memo_id}":{"get":{"tags":["cases"],"summary":"Get Country Conditions Memo","operationId":"get_country_conditions_memo_cases__case_id__country_conditions__memo_id__get","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Case Id"}},{"name":"memo_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Memo Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CountryConditionsMemoDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/webauthn/register/begin":{"post":{"tags":["auth","webauthn"],"summary":"Webauthn Register Begin","operationId":"webauthn_register_begin_auth_webauthn_register_begin_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/webauthn/register/finish":{"post":{"tags":["auth","webauthn"],"summary":"Webauthn Register Finish","operationId":"webauthn_register_finish_auth_webauthn_register_finish_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WebAuthnRegisterFinishRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/webauthn/authenticate/begin":{"post":{"tags":["auth","webauthn"],"summary":"Webauthn Authenticate Begin","operationId":"webauthn_authenticate_begin_auth_webauthn_authenticate_begin_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/webauthn/authenticate/finish":{"post":{"tags":["auth","webauthn"],"summary":"Webauthn Authenticate Finish","operationId":"webauthn_authenticate_finish_auth_webauthn_authenticate_finish_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WebAuthnAuthenticateFinishRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/magic-link":{"post":{"tags":["auth"],"summary":"Request Magic Link","operationId":"request_magic_link_auth_magic_link_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MagicLinkRequest"}}},"required":true},"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/callback":{"get":{"tags":["auth"],"summary":"Magic Link Callback","operationId":"magic_link_callback_auth_callback_get","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string","title":"Token"}}],"responses":{"302":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/logout":{"post":{"tags":["auth"],"summary":"Logout","operationId":"logout_auth_logout_post","responses":{"204":{"description":"Successful Response"}}}},"/auth/me":{"get":{"tags":["auth"],"summary":"Me","operationId":"me_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MeResponse"}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Health Get"}}}}}}}},"components":{"schemas":{"AssignmentChangeEntry":{"properties":{"user_id":{"type":"string","format":"uuid","title":"User Id"},"role_on_case":{"$ref":"#/components/schemas/CaseAssignmentRole"}},"additionalProperties":false,"type":"object","required":["user_id","role_on_case"],"title":"AssignmentChangeEntry"},"AssignmentEntry":{"properties":{"user_id":{"type":"string","format":"uuid","title":"User Id"},"role_on_case":{"$ref":"#/components/schemas/CaseAssignmentRole"}},"additionalProperties":false,"type":"object","required":["user_id","role_on_case"],"title":"AssignmentEntry"},"AsylumOffice":{"type":"string","enum":["arlington","atlanta","boston","chicago","houston","los_angeles","miami","newark","new_york","new_orleans","philadelphia","san_francisco","seattle"],"title":"AsylumOffice"},"CaseAssignmentOut":{"properties":{"user_id":{"type":"string","format":"uuid","title":"User Id"},"role_on_case":{"$ref":"#/components/schemas/CaseAssignmentRole"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["user_id","role_on_case","created_at"],"title":"CaseAssignmentOut"},"CaseAssignmentRole":{"type":"string","enum":["lead_attorney","supporting_attorney","paralegal","supervised_student"],"title":"CaseAssignmentRole"},"CaseAssignmentsPatch":{"properties":{"add":{"items":{"$ref":"#/components/schemas/AssignmentChangeEntry"},"type":"array","maxItems":64,"title":"Add"},"remove":{"items":{"$ref":"#/components/schemas/AssignmentChangeEntry"},"type":"array","maxItems":64,"title":"Remove"}},"additionalProperties":false,"type":"object","title":"CaseAssignmentsPatch"},"CaseCreate":{"properties":{"pseudonym":{"type":"string","maxLength":512,"minLength":1,"title":"Pseudonym"},"country_code":{"type":"string","maxLength":2,"minLength":2,"title":"Country Code"},"basis":{"$ref":"#/components/schemas/ClaimBasis"},"group_description":{"type":"string","maxLength":16384,"minLength":0,"title":"Group Description"},"filing_deadline":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Filing Deadline"},"asylum_office":{"anyOf":[{"$ref":"#/components/schemas/AsylumOffice"},{"type":"null"}]},"intake_notes":{"type":"string","maxLength":16384,"minLength":0,"title":"Intake Notes"},"assignments":{"items":{"$ref":"#/components/schemas/AssignmentEntry"},"type":"array","maxItems":50,"minItems":1,"title":"Assignments"}},"additionalProperties":false,"type":"object","required":["pseudonym","country_code","basis","group_description","intake_notes","assignments"],"title":"CaseCreate"},"CaseDetail":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"pseudonym":{"type":"string","title":"Pseudonym"},"country_code":{"type":"string","title":"Country Code"},"basis":{"$ref":"#/components/schemas/ClaimBasis"},"group_description":{"type":"string","title":"Group Description"},"filing_deadline":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Filing Deadline"},"asylum_office":{"anyOf":[{"$ref":"#/components/schemas/AsylumOffice"},{"type":"null"}]},"intake_notes":{"type":"string","title":"Intake Notes"},"created_by_user_id":{"type":"string","format":"uuid","title":"Created By User Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"archived_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Archived At"},"deleted_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Deleted At"},"assignments":{"items":{"$ref":"#/components/schemas/CaseAssignmentOut"},"type":"array","title":"Assignments"},"can_edit":{"type":"boolean","title":"Can Edit"}},"additionalProperties":false,"type":"object","required":["id","pseudonym","country_code","basis","group_description","filing_deadline","asylum_office","intake_notes","created_by_user_id","created_at","archived_at","deleted_at","assignments","can_edit"],"title":"CaseDetail"},"CaseListItem":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"pseudonym":{"type":"string","title":"Pseudonym"},"country_code":{"type":"string","title":"Country Code"},"basis":{"$ref":"#/components/schemas/ClaimBasis"},"filing_deadline":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Filing Deadline"},"archived_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Archived At"},"deleted_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Deleted At"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"access":{"type":"string","enum":["full","read_only"],"title":"Access"},"assignments":{"items":{"$ref":"#/components/schemas/CaseAssignmentOut"},"type":"array","title":"Assignments"}},"additionalProperties":false,"type":"object","required":["id","pseudonym","country_code","basis","filing_deadline","archived_at","deleted_at","created_at","access","assignments"],"title":"CaseListItem"},"CaseQueryStatus":{"type":"string","enum":["active","archived","deleted","all"],"title":"CaseQueryStatus"},"CaseUpdate":{"properties":{"pseudonym":{"anyOf":[{"type":"string","maxLength":512,"minLength":1},{"type":"null"}],"title":"Pseudonym"},"country_code":{"anyOf":[{"type":"string","maxLength":2,"minLength":2},{"type":"null"}],"title":"Country Code"},"basis":{"anyOf":[{"$ref":"#/components/schemas/ClaimBasis"},{"type":"null"}]},"group_description":{"anyOf":[{"type":"string","maxLength":16384,"minLength":0},{"type":"null"}],"title":"Group Description"},"filing_deadline":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Filing Deadline"},"asylum_office":{"anyOf":[{"$ref":"#/components/schemas/AsylumOffice"},{"type":"null"}]},"intake_notes":{"anyOf":[{"type":"string","maxLength":16384,"minLength":0},{"type":"null"}],"title":"Intake Notes"}},"additionalProperties":false,"type":"object","title":"CaseUpdate"},"ClaimBasis":{"type":"string","enum":["political_opinion","religion","particular_social_group","gender_based","race","nationality","mixed"],"title":"ClaimBasis"},"CountryConditionsGenerateRequest":{"properties":{"country_code":{"type":"string","maxLength":2,"minLength":2,"title":"Country Code"},"basis":{"$ref":"#/components/schemas/ClaimBasis"},"group_description":{"type":"string","maxLength":16384,"minLength":1,"title":"Group Description"},"timeframe_start_year":{"type":"integer","maximum":2100.0,"minimum":1990.0,"title":"Timeframe Start Year"},"jurisdiction_asylum_office":{"anyOf":[{"$ref":"#/components/schemas/AsylumOffice"},{"type":"null"}]}},"additionalProperties":false,"type":"object","required":["country_code","basis","group_description","timeframe_start_year"],"title":"CountryConditionsGenerateRequest"},"CountryConditionsGenerateResponse":{"properties":{"memo_id":{"type":"string","format":"uuid","title":"Memo Id"},"version":{"type":"integer","title":"Version"},"status":{"$ref":"#/components/schemas/CountryConditionsMemoStatus"}},"additionalProperties":false,"type":"object","required":["memo_id","version","status"],"title":"CountryConditionsGenerateResponse"},"CountryConditionsInputs":{"properties":{"country_code":{"type":"string","maxLength":2,"minLength":2,"title":"Country Code"},"basis":{"$ref":"#/components/schemas/ClaimBasis"},"group_description":{"type":"string","maxLength":16384,"minLength":1,"title":"Group Description"},"timeframe_start_year":{"type":"integer","maximum":2100.0,"minimum":1990.0,"title":"Timeframe Start Year"},"jurisdiction_asylum_office":{"anyOf":[{"$ref":"#/components/schemas/AsylumOffice"},{"type":"null"}]}},"additionalProperties":false,"type":"object","required":["country_code","basis","group_description","timeframe_start_year"],"title":"CountryConditionsInputs","description":"Normalized inputs stored on the memo row and passed through the graph."},"CountryConditionsMemoDetail":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"case_id":{"type":"string","format":"uuid","title":"Case Id"},"version":{"type":"integer","title":"Version"},"status":{"$ref":"#/components/schemas/CountryConditionsMemoStatus"},"inputs":{"$ref":"#/components/schemas/CountryConditionsInputs"},"output":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Output"},"model_versions":{"type":"object","title":"Model Versions"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"},"generated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Generated At"}},"additionalProperties":false,"type":"object","required":["id","case_id","version","status","inputs","output","model_versions","error_message"],"title":"CountryConditionsMemoDetail"},"CountryConditionsMemoStatus":{"type":"string","enum":["pending","generating","complete","failed"],"title":"CountryConditionsMemoStatus"},"CountryConditionsMemoSummary":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"case_id":{"type":"string","format":"uuid","title":"Case Id"},"version":{"type":"integer","title":"Version"},"status":{"$ref":"#/components/schemas/CountryConditionsMemoStatus"},"generated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Generated At"}},"additionalProperties":false,"type":"object","required":["id","case_id","version","status"],"title":"CountryConditionsMemoSummary"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MagicLinkRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"organization_slug":{"type":"string","maxLength":128,"minLength":1,"title":"Organization Slug"}},"type":"object","required":["email","organization_slug"],"title":"MagicLinkRequest"},"MeResponse":{"properties":{"user":{"$ref":"#/components/schemas/UserSummary"},"organization":{"$ref":"#/components/schemas/OrganizationSummary"},"mfa_verified":{"type":"boolean","title":"Mfa Verified"},"webauthn_credential_count":{"type":"integer","title":"Webauthn Credential Count"}},"type":"object","required":["user","organization","mfa_verified","webauthn_credential_count"],"title":"MeResponse"},"OrganizationSummary":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name"},"slug":{"type":"string","title":"Slug"}},"type":"object","required":["id","name","slug"],"title":"OrganizationSummary"},"UserRole":{"type":"string","enum":["admin","attorney","paralegal","student"],"title":"UserRole"},"UserSummary":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"email":{"type":"string","title":"Email"},"display_name":{"type":"string","title":"Display Name"},"role":{"$ref":"#/components/schemas/UserRole"}},"type":"object","required":["id","email","display_name","role"],"title":"UserSummary"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WebAuthnAuthenticateFinishRequest":{"properties":{"credential":{"type":"object","title":"Credential"}},"type":"object","required":["credential"],"title":"WebAuthnAuthenticateFinishRequest"},"WebAuthnRegisterFinishRequest":{"properties":{"friendly_name":{"type":"string","maxLength":64,"minLength":1,"title":"Friendly Name"},"credential":{"type":"object","title":"Credential"}},"type":"object","required":["friendly_name","credential"],"title":"WebAuthnRegisterFinishRequest"}}}}